refactor(statistics): factor uid bucketing stats so I can reuse it in test_data generation
