+++ lib/readers/csv.js
@@ -10,10 +10,10 @@
  * uid,email
  */
 
-const assert = require('assert');
 const EventEmitter = require('events');
 const fs = require('fs');
 const readline = require('n-readlines');
+const { updateBucketStats, calculateBucketDistribution } = require('../../lib/statistics');
 
 const UID_HEX32_RE = /^[a-f0-9]{32}$/;
 const SOURCE_FXA = 'fxa';
@@ -46,6 +46,7 @@ class CSVReader extends EventEmitter {
       update: 0,
       delete: 0,
       ignore: 0,
+      stats: {}
     };
 
     while (fxaLine || salesforceLine) {
@@ -58,7 +59,7 @@ class CSVReader extends EventEmitter {
 
       if (comparison < 0) {
         // entry in FxA DB but not the salesforce DB, add it to salesforce.
-        this._updateBucketStats(fxa.uid);
+        updateBucketStats(fxa.uid);
         this.emit('create', fxa);
         counts.create++;
         // catch up until equal or past the salesforceUid
@@ -66,7 +67,7 @@ class CSVReader extends EventEmitter {
       } else if (! comparison) {
         if (fxa.email.toLowerCase() !== salesforce.email.toLowerCase()) {
           // email has changed, need to update salesforce
-          this._updateBucketStats(fxa.uid);
+          updateBucketStats(fxa.uid);
           this.emit('update', fxa);
           counts.update++;
         } else {
@@ -77,38 +78,15 @@ class CSVReader extends EventEmitter {
         salesforceLine = this._nextValidUid(salesforceLineReader, SOURCE_SF);
       } else {
         // Entry is in the salesforce DB but not the FxA DB, remove it from salesforce.
-        this._updateBucketStats(salesforce.uid);
+        updateBucketStats(fxa.uid);
         this.emit('delete', salesforce);
         counts.delete++;
         salesforceLine = this._nextValidUid(salesforceLineReader, SOURCE_SF);
       }
     }
 
-    counts.stats = this._calculateBucketDistribution();
+    counts.stats = calculateBucketDistribution();
     this.emit('complete', counts);
-  }
-
-  _updateBucketStats (uid) {
-    const bucketKey = uid[0];
-    const keys = Object.keys(this.buckets);
-    assert.ok(keys.includes(bucketKey), 'All bucket keys are a-f0-9')
-    this.buckets[bucketKey] += 1;
-  }
-
-  _calculateBucketDistribution () {
-    const bucketKeys = Object.keys(this.buckets);
-    const sum = bucketKeys.reduce((sum, key) => sum + this.buckets[key], 0);
-    const mean = sum / bucketKeys.length;
-    const reducer = (sum, val) => sum + (this.buckets[val] - mean) ** 2;
-    const sumOfSquares = bucketKeys.reduce(reducer, 0);
-    const variance = sumOfSquares / bucketKeys.length;
-    const stddev = Math.sqrt(variance);
-
-    return {
-      sum: sum.toFixed(2),
-      mean: mean.toFixed(2),
-      stddev: stddev.toFixed(2)
-    };
   }
 
   _nextValidUid (reader, source) {
